+++ bin/benchmark.js
@@ -1,92 +1,147 @@
 #!/usr/bin/env node
-require('colors');
+"use strict";
+let startTime = process.hrtime();
 
-var program  = require('commander'),
-    path     = require('path'),
-    async    = require('async'),
-    fs       = require('fs-extra'),
-    du       = require('du'),
-    dir      = require('node-dir'),
-    modclean = require('modclean'),
-    ModClean = modclean.ModClean,
-    Spinner  = require('cli-spinner').Spinner,
-    Table    = require('cli-table'),
-    humanize = require('humanize'),
-    moment   = require('moment');
-    
-program
-    .version(require('../package.json').version)
-    .description('Benchmarking utility for ModClean')
-    .usage('modclean-benchmark [options]')
-    .option('-n, --patterns [patterns]', 'Patterns type(s) to remove (safe, caution, and/or danger)')
-    .option('-m, --modules [modules]', 'Modules to benchmark sepatated by commas')
-    .option('-c, --no-clean', 'Skip cleanup process')
-    .parse(process.argv);
+const program  = require('commander');
+const path     = require('path');
+const async    = require('async');
+const chalk    = require('chalk');
+const fs       = require('fs-extra');
+const du       = require('du');
+const dir      = require('node-dir');
+const modclean = require('modclean');
+const ModClean = modclean.ModClean;
+const ora      = require('ora');
+const Table    = require('cli-table2');
+const humanize = require('humanize');
+const moment   = require('moment');
+const DevNull  = require('dev-null');
 
-console.log('ModClean Benchmarking Utility'.yellow.bold, "\n");
-var totalTime = process.hrtime();
-
-var bmDir = 'mc-benchmark-'+ Date.now(),
-    patterns = typeof program.patterns === 'string' && program.patterns? program.patterns.split(',') : ['safe'],
-    benchmark = {
-        modules: (!program.modules || typeof program.modules === 'boolean')? 
-                    ['express', 'lodash', 'moment', 'async'] : program.modules.split(','),
-        patterns: [],
-        files: [],
-        before: {},
-        after: {},
-        npm: [],
-        mods: {},
-        times: {
-            total: 0,
-            modclean: 0,
-            npm: 0
-        }
-    },
-    spin;
-
-if(Array.isArray(patterns)) {
-    patterns.forEach(function(pat) {
-        if(modclean.patterns.hasOwnProperty(pat))
-            benchmark.patterns.push(modclean.patterns[pat]);
-    });
+function list(val) {
+    return val.split(',');
 }
 
-if(!benchmark.patterns.length)
-    benchmark.patterns.push(modclean.patterns.safe);
+program
+    .version(require('../package.json').version)
+    .description('Benchmarking utility for ModClean 2.x')
+    .usage('modclean-benchmark [options]')
+    .option('-n, --patterns <patterns>', 'Patterns type(s) to remove (safe, caution, and/or danger)', list)
+    .option('-m, --modules <modules>', 'Modules to benchmark sepatated by commas', list)
+    .option('-a, --additional-patterns <list>', 'Additional glob patterns to search for', list)
+    .option('-I, --ignore <list>', 'Comma separated list of patterns to ignore', list)
+    .option('-s, --case-sensitive', 'Matches are case sensitive')
+    .option('--no-dirs', 'Exclude directories from being removed')
+    .option('--no-dotfiles', 'Exclude dot files from being removed')
+    .option('-k, --keep-empty', 'Keep empty directories')
+    .option('--no-clean', 'Skip cleanup process')
+    .option('--no-log', 'Do not create log file at the end')
+    .parse(process.argv);
 
-async.series([
-    // Benchmark setup
-    function setup(cb) {
-        fs.mkdirp('./'+ bmDir +'/node_modules', function(err) {
+console.log(
+    "\n" + chalk.yellow.bold('MODCLEAN 2.x Benchmarking Utility') + "\n"
+);
+
+let spin = ora('');
+    spin.color = 'cyan';
+
+class ModClean_Benchmark {
+    constructor() {
+        this.testName = 'mc-benchmark-'+ Date.now();
+        this.modules = Array.isArray(program.modules) && program.modules.length? program.modules : ['express', 'lodash', 'moment', 'async'];
+        this.patterns = Array.isArray(program.patterns) && program.patterns.length? program.patterns : ['default:safe'];
+        
+        this.info = {
+            files: [],
+            before: {},
+            after: {},
+            npm: [],
+            mods: {},
+            times: {
+                total: 0,
+                modclean: 0,
+                npm: 0
+            }
+        };
+        
+        this.start();
+    }
+    
+    start() {
+        let self = this;
+        
+        async.series([
+            this.setup.bind(this),
+            this.installModules.bind(this),
+            
+            function benchmarkBefore(cb) {
+                spin.text = chalk.gray('Getting stats...');
+                spin.start();
+                
+                self.getStats((stats) => {
+                    self.info.before = stats;
+                    spin.succeed();
+                    cb();
+                });
+            },
+            
+            this.runModclean.bind(this),
+            
+            function benchmarkAfter(cb) {
+                spin.text = chalk.gray('Getting updated stats...');
+                spin.start();
+                
+                self.getStats((stats) => {
+                    self.info.after = stats;
+                    spin.succeed();
+                    cb();
+                });
+            },
+            
+            this.cleanup.bind(this),
+            this.displayResults.bind(this)
+        ], (err) => {
+            if(err) {
+                if(spin) spin.fail();
+                error(err);
+                return process.exit(1);
+            }
+            
+            process.exit(0);
+        });
+    }
+    
+    setup(cb) {
+        fs.mkdirp(`./${this.testName}/node_modules`, (err) => {
             if(err) return cb('Error while creating directory:\n'+ err);
             
-            log(('Installing modules to ./'+ bmDir +'/node_modules').grey);
-            process.chdir('./'+ bmDir);
+            log(chalk.gray(`Installing modules to ./${this.testName}/node_modules`));
+            process.chdir('./' + this.testName);
             cb();
         });
-    },
+    }
     
-    // Install Modules
-    function installModules(cb) {
-        var npm = require('npm'),
+    installModules(cb) {
+        let npm = require('npm'),
             npmTime = process.hrtime();
-        spin = new Spinner('%s'.cyan.bold + ' Installing Modules, please wait...'.grey);
         
-        npm.load(function(err) {
+        log(chalk.gray('Installing modules'), this.modules.join(', '));
+        
+        spin.text = chalk.gray('Installing Modules...');
+        
+        npm.load({ loglevel: "silent", progress: false, logstream: new DevNull() }, (err) => {
             if(err) return cb('Error while loading NPM:\n' + err);
-            log('Installing modules'.grey, benchmark.modules.join(', '));
             spin.start();
             
-            npm.commands.install(benchmark.modules, function(err, data) {
+            npm.commands.install(this.modules, (err, data) => {
                 console.log("\n");
                 if(err) return cb('Error while installing modules:\n' + err);
-                spin.stop(true);
-                benchmark.times.npm = getTime(npmTime);
-                benchmark.npm = data;
+                spin.succeed();
+                
+                this.info.times.npm = getTime(npmTime);
+                this.info.npm = data;
                 
                 if(Array.isArray(data)) {
-                    var mods = {
+                    let mods = {
                         main: [],
                         deps: []
                     };
@@ -96,81 +151,75 @@ async.series([
                         else mods.deps.push(mod[0]);
                     });
                     
-                    benchmark.mods = mods;
+                    this.info.mods = mods;
                 }
                 cb();
             });
         });
-    },
+    }
     
-    // Get stats before ModClean runs
-    function benchmarkBefore(cb) {
-        spin = new Spinner('%s'.cyan.bold + ' Getting stats...'.grey);
-        
-        getStats(function(stats) {
-            benchmark.before = stats;
-            spin.stop(true);
-            cb();
-        });
-    },
-    
-    function runModClean(cb) {
+    runModclean(cb) {
         console.log("\n");
-        spin = new Spinner('%s'.cyan.bold + ' Running ModClean, please wait...'.grey);
-        var mcTime = process.hrtime();
+        spin.text = chalk.gray('Starting ModClean...');
+        spin.start();
+        
+        let mcTime = process.hrtime();
         
         var MC = new ModClean({
             cwd: process.cwd(),
-            patterns: benchmark.patterns
+            patterns: this.patterns,
+            additionalPatterns: program.additionalPatterns || [],
+            ignorePatterns: program.ignore || [],
+            noDirs: !program.dirs,
+            dotFiles: !!program.dotfiles,
+            removeEmptyDirs: !program.keepEmpty,
+            ignoreCase: !program.caseSensitive,
         });
         
-        MC.on('files', function(files) {
-            log('Found'.grey, files.length.toString().green.bold, 'files/folders to remove.'.grey);
-            benchmark.files = files;
+        MC.on('files', (files) => {
+            spin.succeed();
+            log('Found', chalk.green.bold(files.length), 'files/folders to remove.');
+            
+            this.info.files = files;
+            
+            spin.text = chalk.gray('Running ModClean...');
             spin.start();
         });
         
-        MC.clean(function(err, results) {
+        MC.clean((err, results) => {
             if(err) return cb('Error while running ModClean\n'+ err);
-            spin.stop(true);
-            benchmark.times.modclean = getTime(mcTime);
+            spin.succeed();
+            
+            this.info.times.modclean = getTime(mcTime);
             
             cb();
         });
-    },
+    }
     
-    // Get stats after ModClean runs
-    function benchmarkBefore(cb) {
-        spin = new Spinner('%s'.cyan.bold + ' Getting stats...'.grey);
-        
-        getStats(function(stats) {
-            benchmark.after = stats;
-            spin.stop(true);
-            cb();
-        });
-    },
-    
-    // Cleanup modules
-    function cleanup(cb) {
+    cleanup(cb) {
         process.chdir('../');
-        if(program.noClean) return cb();
+        if(!program.clean) return cb();
         
-        spin = new Spinner('%s'.cyan.bold + ' Cleaning up, please wait...'.grey);
+        spin.text = chalk.gray('Cleaning up...');
         spin.start();
         
-        fs.remove('./'+ bmDir, function(err) {
-            spin.stop(true);
-            if(err) error('Unable to cleanup files installed, please cleanup manually');
+        fs.remove('./'+ this.testName, (err) => {
+            if(err) {
+                spin.fail();
+                error('Unable to cleanup files installed, please cleanup manually');
+                return cb();
+            }
+            
+            spin.succeed();
             cb();
         });
-    },
+    }
     
-    // Display the results
-    function displayResults(cb) {
+    displayResults(cb) {
         console.log("\n");
-        log('Results for', ('npm install '+ benchmark.modules.join(' ')).grey, 'using patterns', patterns.join(', ').grey);
+        log(`Results for ${chalk.gray('npm install '+ this.modules.join(' '))} using patterns ${chalk.gray(this.patterns.join(', '))}`);
         
-        var table = new Table({
+        let table = new Table({
             head: ['', 'Total Files', 'Total Folders', 'Total Size'],
             style: {
                 head: ['yellow', 'bold']
@@ -179,123 +228,111 @@ async.series([
         
         table.push({
             'Before ModClean': [
-                formatNumber(benchmark.before.files),
-                formatNumber(benchmark.before.dirs),
-                humanize.filesize(benchmark.before.size)
+                formatNumber(this.info.before.files),
+                formatNumber(this.info.before.dirs),
+                humanize.filesize(this.info.before.size)
             ]
         }, {
             'After ModClean': [
-                formatNumber(benchmark.after.files),
-                formatNumber(benchmark.after.dirs),
-                humanize.filesize(benchmark.after.size)
+                formatNumber(this.info.after.files),
+                formatNumber(this.info.after.dirs),
+                humanize.filesize(this.info.after.size)
             ]
         }, {
             'Reduced': [
-                formatNumber(benchmark.before.files - benchmark.after.files),
-                formatNumber(benchmark.before.dirs - benchmark.after.dirs),
-                humanize.filesize(benchmark.before.size - benchmark.after.size)
+                formatNumber(this.info.before.files - this.info.after.files),
+                formatNumber(this.info.before.dirs - this.info.after.dirs),
+                humanize.filesize(this.info.before.size - this.info.after.size)
             ]
         });
         
-        var tbl = table.toString();
+        let tbl = table.toString();
         
         console.log(tbl, "\n");
-        benchmark.times.total = getTime(totalTime);
+        this.info.times.total = getTime(startTime);
         
-        console.log('NPM Install Time:'.yellow.bold, formatTime(benchmark.times.npm));
-        console.log('ModClean Time:   '.yellow.bold, formatTime(benchmark.times.modclean));
-        console.log('Total Time:      '.yellow.bold, formatTime(benchmark.times.total));
+        console.log(chalk.yellow.bold('NPM Install Time:'), formatTime(this.info.times.npm));
+        console.log(chalk.yellow.bold('ModClean Time:   '), formatTime(this.info.times.modclean));
+        console.log(chalk.yellow.bold('Total Time:      '), formatTime(this.info.times.total));
         
-        var logTable = new Table({
+        if(!program.log) return cb();
+        
+        let logTable = new Table({
             head: ['', 'Result']
         });
         
         logTable.push({
-            'Modules': [benchmark.modules.join(', ')]
+            'Modules': [this.modules.join(', ')]
         }, {
-            'Installed Modules': [benchmark.mods.main.join(', ')]
+            'Patterns': [this.patterns.join(', ')]
         }, {
-            'Dependencies': [benchmark.mods.deps.join(', ')]
+            'ModClean Removed': [formatNumber(this.info.files.length)]
         }, {
-            'Patterns': [benchmark.patterns.join(', ')]
+            'Total Removed': [formatNumber((this.info.before.files - this.info.after.files) + (this.info.before.dirs - this.info.after.dirs))]
         }, {
-            'ModClean Removed': [formatNumber(benchmark.files.length)]
+            'NPM Install Time': [formatTime(this.info.times.npm)]
         }, {
-            'Total Removed': [formatNumber((benchmark.before.files - benchmark.after.files) + (benchmark.before.dirs - benchmark.after.dirs))]
+            'ModClean Time': [formatTime(this.info.times.modclean)]
         }, {
-            'NPM Install Time': [formatTime(benchmark.times.npm)]
-        }, {
-            'ModClean Time': [formatTime(benchmark.times.modclean)]
-        }, {
-            'Total Time': [formatTime(benchmark.times.total)]
+            'Total Time': [formatTime(this.info.times.total)]
         });
         
-        var logFile = [
+        let logFile = [
             tbl.replace(/\u001b\[(?:\d*;){0,5}\d*m/g, ''),
             "\n",
             logTable.toString().replace(/\u001b\[(?:\d*;){0,5}\d*m/g, ''),
             "\n",
-            'RAW DATA:',
-            '-----------------------------------------------------------------------------------',
-            JSON.stringify(benchmark, null, 4)
+            JSON.stringify(this.info, null, 4)
         ];
         
-        fs.outputFile('./'+ bmDir +'.log', logFile.join("\n"), function(err) {
+        fs.outputFile('./'+ this.testName +'.log', logFile.join("\n"), (err) => {
             console.log("\n");
             if(err) error('Unable to write log file');
-            else log('Log file written to', ('./'+ bmDir +'.log').grey);
+            else log('Log file written to', chalk.gray('./'+ this.testName +'.log'));
             cb();
         });
     }
-], function(err) {
-    if(err) {
-        if(spin) spin.stop(true);
-        error(err);
-        return process.exit(1);
-    }
     
-    process.exit(0);
-});
+    getStats(cb) {
+        let fp = path.join('./node_modules'),
+            stats = {
+                files: 0,
+                dirs: 0,
+                size: 0
+            };
+        
+        dir.files(fp, function(err, files) {
+            if(err) error('Error while getting file stats:\n', err);
+            else stats.files = files.length;
+            
+            dir.subdirs(fp, function(err, dirs) {
+                if(err) error('Error while getting directory stats:\n', err);
+                else stats.dirs = dirs.length;
+                
+                du(fp, function(err, size) {
+                    if(err) error('Error while getting directory stats:\n', err);
+                    else stats.size = size;
+                    
+                    cb(stats);
+                });
+            });
+        });
+    }
+}
+
+new ModClean_Benchmark();
 
 ////////////////////////////////////////////////
 /// UTILITY FUNCTIONS                         //
 ////////////////////////////////////////////////
 
-function getStats(cb) {
-    var fp = path.join('./node_modules'),
-        stats = {
-            files: 0,
-            dirs: 0,
-            size: 0
-        };
-    
-    dir.files(fp, function(err, files) {
-        if(err) error('Error while getting file stats:\n', err);
-        else stats.files = files.length;
-        
-        dir.subdirs(fp, function(err, dirs) {
-            if(err) error('Error while getting directory stats:\n', err);
-            else stats.dirs = dirs.length;
-            
-            du(fp, function(err, size) {
-                if(err) error('Error while getting directory stats:\n', err);
-                else stats.size = size;
-                
-                cb(stats);
-            });
-        });
-    });
-}
-
-function log() {
-    var args = Array.prototype.slice.call(arguments, 0);
-    args.unshift(' \u25BA '.cyan.bold);
+function log(...args) {
+    args.unshift(chalk.cyan.bold(' \u25BA '));
     console.log.apply(console, args);
 }
 
-function error() {
-    var args = Array.prototype.slice.call(arguments, 0);
-    args.unshift(' X '.red.bold);
+function error(...args) {
+    args.unshift(chalk.red.bold(' X '));
     console.error.apply(console, args);
 }
 
@@ -304,14 +341,14 @@ function formatNumber(x) {
 }
 
 function getTime(time) {
-    var diff = process.hrtime(time);
+    let diff = process.hrtime(time);
     return (diff[0] * 1e9 + diff[1]) / 1000000;
 }
 
 function formatTime(secs) {
     if(typeof secs !== 'number') return 'Not Available';
     
-    var d = moment.duration(secs, 'milliseconds'),
+    let d = moment.duration(secs, 'milliseconds'),
         years = d.years(),
         months = d.months(),
         days = d.days(),
@@ -323,8 +360,8 @@ function formatTime(secs) {
     if(years > 0) out.push(years + ' year' + (years>1? 's' : ''));
     if(months > 0) out.push(months + ' month' + (months>1? 's' : ''));
     if(days > 0) out.push(days + ' day' + (days>1? 's' : ''));
+    if(hours > 0) out.push(hours + ' hour' + (hours>1? 's' : ''));
     
-    out.push(hours + ' hour' + (hours<1||hours>1? 's' : ''));
     out.push(mins + ' minute' + (mins<1||mins>1? 's' : ''));
     out.push(seconds + ' second' + (seconds<1||seconds>1? 's' : ''));