+++ HISTORY.md
@@ -1,4 +1,23 @@
 # ModClean Benchmark History
 
+## 2.0.0 (1/11/2016)
+* **New!** Added `-a, --additional-patterns` option.
+* **New!** Added `-I, --ignore` option.
+* **New!** Added `-s, --case-sensitive` option.
+* **New!** Added `--no-dirs` option.
+* **New!** Added `--no-dotfiles` option.
+* **New!** Added `-k, --keep-empty` option.
+* **New!** Added `--no-log` option.
+* **Breaking** Replaced `-d, --empty` option with the new `-k, --keep-empty` option.
+* Rewritten using ES6 (requires Node v6.9+)
+* Updated all dependencies to their latest versions.
+* Updated to ModClean 2.0.0.
+* Replaced `colors` with `chalk`.
+* Replaced `cli-table` with `cli-table2`.
+* Replaced `cli-spinner` with `ora`.
+* Improved output of NPM commands.
+* Reduced log file output.
+* Misc. display and performance updates/fixes.
+
 ## 1.0.0 (4/27/2015)
 * Initial release